+++ packages/core/src/query.ts
@@ -1,5 +1,6 @@
 import { Doc } from './classes'
 import { createPredicates, isPredicate } from './predicate'
+import { SortingQuery } from './storage'
 
 export const likeSymbol = '%'
 
@@ -24,4 +25,15 @@ export function findProperty (objects: Doc[], propertyKey: string, value: any): 
     }
   }
   return result
+}
+
+export function resultSort<T extends Doc> (result: T[], sortOptions: SortingQuery<T>): void {
+  const sortFunc = (a: any, b: any): number => {
+    for (const key in sortOptions) {
+      const result = typeof a[key] === 'string' ? a[key].localeCompare(b[key]) : a - b
+      if (result !== 0) return result * (sortOptions[key] as number)
+    }
+    return 0
+  }
+  result.sort(sortFunc)
 }
